add sys-cluster/modules package
Add Environment Modules package in EESSI Gentoo overlay to be able to
support this module tool (see [1] for context).

This commit adds the ebuild for Modules version 5.6.1 and updates sets
to install this package.

Specific installation options for this package (especially to be able to
install both Lmod and Modules) are introduced by
EESSI/compatibility-layer#231.

[1] https://gitlab.com/eessi/support/-/issues/195

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>